+++ src/coverlet.collector/DataCollection/CoverageWrapper.cs
@@ -15,7 +15,7 @@ namespace Coverlet.Collector.DataCollection
         /// <param name="settings">Coverlet settings</param>
         /// <param name="coverletLogger">Coverlet logger</param>
         /// <returns>Coverage object</returns>
-        public Coverage CreateCoverage(CoverletSettings settings, ILogger coverletLogger, IInstrumentationHelper instrumentationHelper, IFileSystem fileSystem)
+        public Coverage CreateCoverage(CoverletSettings settings, ILogger coverletLogger, IInstrumentationHelper instrumentationHelper, IFileSystem fileSystem, ISourceRootTranslator sourceRootTranslator)
         {
             return new Coverage(
                 settings.TestModule,
@@ -30,7 +30,8 @@ namespace Coverlet.Collector.DataCollection
                 settings.UseSourceLink,
                 coverletLogger,
                 instrumentationHelper,
-                fileSystem);
+                fileSystem,
+                sourceRootTranslator);
         }
 
         /// <summary>